feat: Add support for displaying and updating project events
This adds support for displaying and calendar. It introduces a new configuration option, `display.projects.display`, which controls whether project events should be displayed or not. If set to `true`, project events will be created or updated based on the project data retrieved from the API. If set to `false`, project events will be deleted if they already calendar.